+++ src/qqtools/plugins/qpipeline/runner/runner_utils/ddpdeduper/__init__.py
@@ -4,7 +4,13 @@ from .eval_contract import (
     EvalDedupRuntime,
     unwrap_eval_batch,
 )
-from .output_deduper import DDPOutputDeduper, wrap_eval_dataloader_for_ddp_dedup
+from .output_deduper import (
+    DDPOutputDeduper,
+    prepare_eval_loader_for_ddp,
+    resolve_loader_is_graph,
+    validate_eval_loader_batch_counts,
+    wrap_eval_dataloader_for_ddp_dedup,
+)
 
 __all__ = [
     "DDPOutputDeduper",
@@ -13,4 +19,7 @@ __all__ = [
     "EvalDedupRuntime",
     "unwrap_eval_batch",
     "wrap_eval_dataloader_for_ddp_dedup",
+    "prepare_eval_loader_for_ddp",
+    "resolve_loader_is_graph",
+    "validate_eval_loader_batch_counts",
 ]